+++ src/pieces/king.rs
@@ -164,779 +164,3 @@ impl King {
         valid_for_castling
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        board::{Board, Coord},
-        pieces::{king::King, PieceColor, PieceMove, PieceType, Position},
-        utils::is_getting_checked,
-    };
-
-    #[test]
-    fn multiple_enemies_1() {
-        let custom_board = [
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::King, PieceColor::White)),
-                None,
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(4, 5), Coord::new(5, 4)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(4, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn multiple_enemies_2() {
-        let custom_board = [
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                Some((PieceType::Queen, PieceColor::Black)),
-                None,
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::King, PieceColor::White)),
-                None,
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(3, 4)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(4, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn multiple_enemies_3() {
-        let custom_board = [
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-                None,
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-            [
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                None,
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-            ],
-            [
-                None,
-                None,
-                Some((PieceType::Bishop, PieceColor::Black)),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(4, 5)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(4, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn big_castle_white() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                None,
-                None,
-                None,
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(7, 3), Coord::new(7, 0)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn small_castle_white() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(7, 5), Coord::new(7, 7)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn big_castle_black() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                None,
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(7, 4), Coord::new(7, 7)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 3),
-            PieceColor::Black,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-    #[test]
-    fn small_castle_black() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                None,
-                None,
-                Some((PieceType::King, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(7, 0), Coord::new(7, 2)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 3),
-            PieceColor::Black,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn big_castle_black_check_blocking() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                None,
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                None,
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-                None,
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(0, 5)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(0, 4),
-            PieceColor::Black,
-            board.board,
-            &[],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn big_castle_black_king_checked() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                None,
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                None,
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                None,
-                Some((PieceType::King, PieceColor::White)),
-            ],
-        ];
-        let mut board = Board::new(custom_board, PieceColor::Black, vec![]);
-        board.set_board(custom_board);
-
-        let is_king_checked =
-            is_getting_checked(board.board, board.player_turn, &board.move_history);
-
-        let mut right_positions = vec![Coord::new(0, 5)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(0, 4),
-            PieceColor::Black,
-            board.board,
-            &[],
-            is_king_checked,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn big_castle_black_rook_already_moved() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Queen, PieceColor::Black)),
-                Some((PieceType::King, PieceColor::Black)),
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                None,
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-        ];
-        let mut board = Board::default();
-        board.set_board(custom_board);
-
-        let mut right_positions = vec![Coord::new(0, 5)];
-        right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(0, 4),
-            PieceColor::Black,
-            board.board,
-            &[
-                (PieceMove {
-                    piece_type: PieceType::Rook,
-                    piece_color: PieceColor::White,
-                    from: Coord::new(0, 7),
-                    to: Coord::new(4, 7),
-                }),
-                (PieceMove {
-                    piece_type: PieceType::Pawn,
-                    piece_color: PieceColor::Black,
-                    from: Coord::new(6, 2),
-                    to: Coord::new(5, 2),
-                }),
-                (PieceMove {
-                    piece_type: PieceType::Rook,
-                    piece_color: PieceColor::White,
-                    from: Coord::new(4, 7),
-                    to: Coord::new(0, 7),
-                }),
-            ],
-            false,
-        );
-        positions.sort();
-
-        assert_eq!(right_positions, positions);
-    }
-
-    #[test]
-    fn castle_both_sides() {
-        let custom_board = [
-            [
-                Some((PieceType::Rook, PieceColor::Black)),
-                None,
-                None,
-                None,
-                Some((PieceType::King, PieceColor::Black)),
-                Some((PieceType::Bishop, PieceColor::Black)),
-                Some((PieceType::Knight, PieceColor::Black)),
-                Some((PieceType::Rook, PieceColor::Black)),
-            ],
-            [
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-                Some((PieceType::Pawn, PieceColor::Black)),
-            ],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [None, None, None, None, None, None, None, None],
-            [
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-                Some((PieceType::Pawn, PieceColor::White)),
-            ],
-            [
-                Some((PieceType::Rook, PieceColor::White)),
-                Some((PieceType::Knight, PieceColor::White)),
-                Some((PieceType::Bishop, PieceColor::White)),
-                Some((PieceType::Queen, PieceColor::White)),
-                Some((PieceType::King, PieceColor::White)),
-                None,
-                None,
-                Some((PieceType::Rook, PieceColor::White)),
-            ],
-        ];
-        let mut board = Board::new(custom_board, PieceColor::White, vec![]);
-        board.set_board(custom_board);
-
-        let is_king_checked =
-            is_getting_checked(board.board, board.player_turn, &board.move_history);
-
-        let mut white_right_positions = vec![Coord::new(7, 5), Coord::new(7, 7)];
-        white_right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 4),
-            PieceColor::White,
-            board.board,
-            &[],
-            is_king_checked,
-        );
-        positions.sort();
-
-        assert_eq!(white_right_positions, positions);
-
-        board.move_piece_on_the_board(&Coord::new(7, 4), &Coord::new(7, 7));
-        board.flip_the_board();
-
-        let mut black_right_positions = vec![Coord::new(7, 4), Coord::new(7, 7)];
-        black_right_positions.sort();
-
-        let mut positions = King::authorized_positions(
-            &Coord::new(7, 3),
-            PieceColor::Black,
-            board.board,
-            &[],
-            is_king_checked,
-        );
-        positions.sort();
-
-        assert_eq!(black_right_positions, positions);
-    }
-}
