+++ tests/pieces/knight.rs
@@ -1,0 +1,291 @@
+#[cfg(test)]
+mod tests {
+    use chess_tui::board::{Board, Coord};
+    use chess_tui::pieces::knight::Knight;
+    use chess_tui::pieces::{PieceColor, PieceType, Position};
+    use chess_tui::utils::is_getting_checked;
+
+    #[test]
+    fn no_enemies() {
+        let custom_board = [
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::White)),
+                None,
+                None,
+                None,
+            ],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+        ];
+        let mut board = Board::default();
+        board.set_board(custom_board);
+
+        let mut right_positions = vec![
+            Coord::new(2, 3),
+            Coord::new(2, 5),
+            Coord::new(3, 2),
+            Coord::new(3, 6),
+            Coord::new(5, 2),
+            Coord::new(5, 6),
+            Coord::new(6, 3),
+            Coord::new(6, 5),
+        ];
+        right_positions.sort();
+
+        let mut positions = Knight::authorized_positions(
+            &Coord::new(4, 4),
+            PieceColor::White,
+            board.board,
+            &[],
+            false,
+        );
+        positions.sort();
+
+        assert_eq!(right_positions, positions);
+    }
+
+    #[test]
+    fn enemy_and_ally() {
+        let custom_board = [
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Pawn, PieceColor::White)),
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Pawn, PieceColor::Black)),
+                None,
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::White)),
+            ],
+        ];
+        let mut board = Board::default();
+        board.set_board(custom_board);
+
+        let mut right_positions = vec![Coord::new(6, 5)];
+        right_positions.sort();
+
+        let mut positions = Knight::authorized_positions(
+            &Coord::new(7, 7),
+            PieceColor::White,
+            board.board,
+            &[],
+            false,
+        );
+        positions.sort();
+
+        assert_eq!(right_positions, positions);
+    }
+
+    #[test]
+    fn king_checked_can_resolve() {
+        let custom_board = [
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::King, PieceColor::White)),
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::White)),
+                None,
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::Black)),
+            ],
+        ];
+        let mut board = Board::new(custom_board, PieceColor::White, vec![]);
+        board.set_board(custom_board);
+
+        let is_king_checked =
+            is_getting_checked(board.board, board.player_turn, &board.move_history);
+
+        let mut right_positions = vec![Coord::new(7, 7)];
+        right_positions.sort();
+
+        let mut positions = Knight::authorized_positions(
+            &Coord::new(6, 5),
+            PieceColor::White,
+            board.board,
+            &[],
+            is_king_checked,
+        );
+        positions.sort();
+
+        assert_eq!(right_positions, positions);
+    }
+
+    #[test]
+    fn king_checked_cant_resolve() {
+        let custom_board = [
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::King, PieceColor::White)),
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::White)),
+                None,
+                None,
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::Black)),
+            ],
+        ];
+        let mut board = Board::new(custom_board, PieceColor::White, vec![]);
+        board.set_board(custom_board);
+
+        let is_king_checked =
+            is_getting_checked(board.board, board.player_turn, &board.move_history);
+
+        let mut right_positions: Vec<Coord> = vec![];
+        right_positions.sort();
+
+        let mut positions = Knight::authorized_positions(
+            &Coord::new(6, 4),
+            PieceColor::White,
+            board.board,
+            &[],
+            is_king_checked,
+        );
+        positions.sort();
+
+        assert_eq!(right_positions, positions);
+    }
+    #[test]
+    fn nailing() {
+        let custom_board = [
+            [
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::King, PieceColor::Black)),
+                None,
+                None,
+                None,
+            ],
+            [
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Knight, PieceColor::Black)),
+                None,
+                None,
+                None,
+            ],
+            [None, None, None, None, None, None, None, None],
+            [
+                None,
+                None,
+                None,
+                None,
+                Some((PieceType::Queen, PieceColor::White)),
+                None,
+                None,
+                None,
+            ],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+            [None, None, None, None, None, None, None, None],
+        ];
+        let mut board = Board::new(custom_board, PieceColor::Black, vec![]);
+        board.set_board(custom_board);
+
+        let is_king_checked =
+            is_getting_checked(board.board, board.player_turn, &board.move_history);
+
+        let mut right_positions: Vec<Coord> = vec![];
+        right_positions.sort();
+
+        let mut positions = Knight::authorized_positions(
+            &Coord::new(1, 4),
+            PieceColor::Black,
+            board.board,
+            &[],
+            is_king_checked,
+        );
+        positions.sort();
+
+        assert_eq!(right_positions, positions);
+    }
+}
