+++ tests/pieces/mod.rs
@@ -1,0 +1,6 @@
+mod bishop;
+mod king;
+mod knight;
+mod pawn;
+mod queen;
+mod rook;
